+++ examples/main/modules.tf
@@ -15,20 +15,8 @@ module "rg" {
   stack       = var.stack
 }
 
-module "run" {
-  source  = "claranet/run/azurerm"
-  version = "x.x.x"
-
-  client_name         = var.client_name
-  environment         = var.environment
-  stack               = var.stack
-  location            = module.azure_region.location
-  location_short      = module.azure_region.location_short
-  resource_group_name = module.rg.resource_group_name
-}
-
-module "sentinel" {
-  source  = "claranet/sentinel/azurerm"
+module "logs" {
+  source  = "claranet/run/azurerm//modules/logs"
   version = "x.x.x"
 
   location            = module.azure_region.location
@@ -38,13 +26,12 @@ module "sentinel" {
   client_name = var.client_name
   environment = var.environment
   stack       = var.stack
+}
 
-  logs_destinations_ids = [
-    module.run.logs_storage_account_id,
-    module.run.log_analytics_workspace_id
-  ]
+module "sentinel" {
+  source  = "claranet/sentinel/azurerm"
+  version = "x.x.x"
 
-  extra_tags = {
-    foo = "bar"
-  }
+  log_analytics_workspace_id = module.logs.log_analytics_workspace_id
+  data_connector_aad_enabled = true
 }