+++ r-sentinel.tf
@@ -1,8 +1,3 @@
-resource "azurerm_sentinel_data_connector_azure_security_center" "sentinel" {
-  name = local.sentinel_name
-
-  location            = var.location
-  resource_group_name = var.resource_group_name
-
-  tags = merge(local.default_tags, var.extra_tags)
+resource "azurerm_sentinel_log_analytics_workspace_onboarding" "sentinel" {
+  workspace_id = var.log_analytics_workspace_id
 }