+++ variables.tf
@@ -1,30 +1,34 @@
-#### Common variables
-variable "resource_group_name" {
-  description = "Name of the resource group."
+variable "log_analytics_workspace_id" {
+  description = "The Log Analytics Workspace ID."
   type        = string
 }
 
-variable "location" {
-  description = "Azure region to use."
-  type        = string
+variable "data_connector_aad_enabled" {
+  description = "Whether the Azure Active Directory logs are retrieved."
+  type        = bool
+  default     = false
 }
 
-variable "location_short" {
-  description = "Short string for Azure location."
-  type        = string
+variable "data_connector_aad_logs" {
+  description = "List of Azure Active Directory log category."
+  type        = list(string)
+  default     = ["AuditLogs", "SignInLogs", "NonInteractiveUserSignInLogs", "ServicePrincipalSignInLogs", "ManagedIdentitySignInLogs", "ProvisioningLogs", "ADFSSignInLogs", "RiskyUsers", "UserRiskEvents", "NetworkAccessTrafficLogs", "RiskyServicePrincipals", "ServicePrincipalRiskEvents", "EnrichedOffice365AuditLogs", "MicrosoftGraphActivityLogs"]
 }
 
-variable "client_name" {
-  description = "Client name/account used in naming."
-  type        = string
+variable "ueba_enabled" {
+  description = "Whether UEBA (User and Entity Behavior Analytics) feature is enabled."
+  type        = bool
+  default     = true
 }
 
-variable "environment" {
-  description = "Project environment."
-  type        = string
+variable "ueba_data_sources" {
+  description = "List of UEBA (User and Entity Behavior Analytics) data sources."
+  type        = list(string)
+  default     = ["AuditLogs", "AzureActivity", "SecurityEvent", "SigninLogs"]
 }
 
-variable "stack" {
-  description = "Project stack name."
-  type        = string
+variable "ueba_entity_providers" {
+  description = "List of UEBA (User and Entity Behavior Analytics) entity providers."
+  type        = list(string)
+  default     = ["AzureActiveDirectory"]
 }