+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Unreleased
 
 Added
-  * AZ-1320: Azure Sentinel module first release
+  * AZ-1320: Microsoft Sentinel module first release